// types/type.ts
export type Attributes = {
  width: string
  height: string
  fontSize: string
  fontFamily: string
  fontWeight: string
  fill: string
  stroke: string
}
// ...
export type RightSidebarProps = {
  elementAttributes: Attributes
  setElementAttributes: React.Dispatch<React.SetStateAction<Attributes>>
  fabricRef: React.RefObject<fabric.Canvas | null>
  activeObjectRef: React.RefObject<fabric.Object | null>
  isEditingRef: React.MutableRefObject<boolean>
  syncShapeInStorage: (obj: any) => void
}